# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Risk APIs",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/risk/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Risk"
                ],
                "summary": "get a risk by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "risk uuid",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Risk"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/risks": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Risk"
                ],
                "summary": "get a list of risks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Risk"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Risk"
                ],
                "summary": "create a risk",
                "parameters": [
                    {
                        "description": "a new risk (without id)",
                        "name": "risk",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Risk"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new created risk with uuid",
                        "schema": {
                            "$ref": "#/definitions/models.Risk"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Risk": {
            "type": "object",
            "required": [
                "state",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "id": {
                    "type": "string"
                },
                "state": {
                    "type": "string",
                    "enum": [
                        "open",
                        "closed",
                        "accepted",
                        "investigating"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        }
    }
}